+++ leoz-central/src/main/java/org/deku/leoz/central/service/internal/OrderService.kt
@@ -67,6 +67,7 @@ class OrderService : OrderService {
                     authorizedUserRecord ?:
                             throw DefaultProblem(status = Response.Status.UNAUTHORIZED)
                     when {
+//todo include send_date
                         order.findOrderDebitorId(Type.DELIVERY) != authorizedUserRecord.debitorId
                         ->
                             throw DefaultProblem(status = Response.Status.FORBIDDEN)
@@ -97,7 +98,7 @@ class OrderService : OrderService {
                 ?: throw DefaultProblem(
                 title = "Order not found",
                 status = Response.Status.NOT_FOUND)
-//to be removed if mobile supports apikeys
+//todo to be removed if mobile supports apikeys
         if (apiKey != null) {
 //--<
             apiKey ?:
@@ -107,6 +108,7 @@ class OrderService : OrderService {
             authorizedUserRecord ?:
                     throw DefaultProblem(status = Response.Status.UNAUTHORIZED)
             when {
+//todo include send_date
                 order.findOrderDebitorId(Type.DELIVERY) != authorizedUserRecord.debitorId
                 ->
                     throw DefaultProblem(status = Response.Status.FORBIDDEN)
@@ -135,9 +137,9 @@ class OrderService : OrderService {
             }
         }
     }
-
+//todo include send_date
     fun Order.findOrderDebitorId(pickupDeliver: Type): Int {
-        var station: Long
+        var station: Int
 
         when (pickupDeliver) {
             Type.PICKUP -> station = this.pickupStation
@@ -167,7 +169,7 @@ class OrderService : OrderService {
             o.orderClassification = OrderClassification.PICKUP_DELIVERY
         else
             o.orderClassification = OrderClassification.DELIVERY
-        o.pickupStation = r.pickupStation.toLong()
+        o.pickupStation = r.pickupStation.toInt()
         o.pickupAddress.line1 = r.pickupAddressLine1
         o.pickupAddress.line2 = r.pickupAddressLine2
         o.pickupAddress.line3 = r.pickupAddressLine3
@@ -203,7 +205,7 @@ class OrderService : OrderService {
         o.pickupAppointment.dateEnd = r.appointmentPickupEnd
         o.pickupAppointment.notBeforeStart = r.appointmentPickupNotBeforeStart == 1
 
-        o.deliveryStation = r.deliveryStation.toLong()
+        o.deliveryStation = r.deliveryStation.toInt()
         o.deliveryAddress.line1 = r.deliveryAddressLine1
         o.deliveryAddress.line2 = r.deliveryAddressLine2
         o.deliveryAddress.line3 = r.deliveryAddressLine3
